requirements.txt
# Core dependencies
fastapi>=0.104.0
passlib>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
typing-extensions>=4.7.0

# Password hashing backend
bcrypt<4.0.0

# Security
cryptography>=41.0.0
pyjwt>=2.8.0

# Database
sqlalchemy>=2.0.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0

# ASGI server
uvicorn[standard]>=0.23.2
